requirements.txt
pdfplumber==0.11.6
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-genai>=1.0.0
